refactor: remove synchronous exec (#47)

+++ routes/main/get.js
@@ -1,7 +1,8 @@
 const { promises: fs } = require('fs')
 const path = require('path')
-const cp = require('child_process')
 const os = require('os')
+
+const asyncExec = require('../../helpers/asyncExec.js')
 
 module.exports = {
   method: 'GET',
@@ -25,7 +26,8 @@ module.exports = {
     const {
       getJobData,
       defaultTemplate,
-      templatesDir
+      templatesDir,
+      formatPrntCommandData
     } = this.cloudPrnt
 
     const jobData = await getJobData(token)
@@ -36,20 +38,21 @@ module.exports = {
     const template = jobData.template || defaultTemplate
     const templatePath = path.join(templatesDir, template)
     const renderedJob = await this.view(templatePath, jobData)
-
-    const printFilePath = path.join(os.tmpdir(), `${token}.stm`)
-    await fs.writeFile(printFilePath, renderedJob, { flag: 'w+' })
-
     const mimeType = 'application/vnd.star.starprntcore'
-    const prntCommandData = cp.execSync(
-        `cputil decode ${mimeType} ${printFilePath} -`,
-        {
-          encoding: null,
-          stdio: ['ignore', 'pipe', 'pipe']
-        }
-    )
 
-    await fs.unlink(printFilePath)
+    let prntCommandData
+
+    if (typeof formatPrntCommandData === 'function') {
+      prntCommandData = await formatPrntCommandData(renderedJob)
+    } else {
+      const printFilePath = path.join(os.tmpdir(), `${token}.stm`)
+      await fs.writeFile(printFilePath, renderedJob, { flag: 'w+' })
+      prntCommandData = await asyncExec(
+        'cputil',
+        ['decode', mimeType, printFilePath, '-']
+      )
+      await fs.unlink(printFilePath)
+    }
 
     return reply
       .type(mimeType)
